* Changed the App package name * Changed the settings Fragment to a PreferenceActivity * Added an IntegerListPreference preference option * Changed the SharedPreferenceKey references to a more explaining name * Eclipse project cleanup

+++ src/io/github/thewear/thewearandroidClientAPP/PreferenceConvertor.java
@@ -77,8 +77,8 @@ public class PreferenceConvertor {
 			case 3: // Beaufort, conversion needed
 				prefMax = SettingsConvertor.metersToBeaufort(res
 						.getInteger(maxPreferenceIdentifier));
-				pref = (int) (prefMax - SettingsConvertor.metersToBeaufort((res
-						.getInteger(minPreferenceIdentifier))));
+				pref = prefMax - SettingsConvertor.metersToBeaufort((res
+						.getInteger(minPreferenceIdentifier)));
 				break;
 			case 4: // Knots, conversion needed
 				prefMax = (int) SettingsConvertor.metersToKnots(res
